Rename 28-Find The Nearest Supporter to 29-Find The Nearest Supporter
There are contestants standing in a row. Each contestant is assigned a rating, which is an integer. A contestant supports all those whose rating is greater than their rating.

Now for each contestant standing in the row, you need to tell the rating of its closest supporter on the left.

If no supporter is there, store -1 in that place.

Input Format
The first line an integer n which denotes the size of the row array.

Next line contains n space separated integers which are the elements of the row array.

Output Format
Return an integer array in which each element denotes the rating of the closest supporter on the left.

(https://www.codingninjas.com/studio/problems/find-the-nearest-supporter_1082335?leftPanelTab=0)

Example 1
Input

5
3 1 5 12 10
Output

-1 -1 1 5 5
Explanation

There is no contestant on the left of 1st contestant
No contestant’s rating is less than this player, hence -1.
Both ratings 3 and 1 are smaller than the current rating, but ‘1’ is closest.
All ratings are small, but the nearest one is ‘5’.
For ‘10’, the nearest rating is ‘5’, as contestant 4 is not a supporter.

Example 2
Input

6
9 8 1 2 5 5
Output

-1 -1 -1 1 2 2
Explanation

For the first ''3'' contestants, there is no supporter.
For contestant ''4'', the only supporter is contestant ''3'', with the rating ''1''.
Both ratings ''1'' and ''2'' are smaller than the current rating, but ‘2’ is closest.
The rating of contestant ''5'' is not less than the current contestant, hence the closest supporter is 4th, with a rating of ''2''.

Constraints

1 <= n <= 10^5

1 <= row[i] <= 10^3